Add debug logging of bytes sent/received (#10)
* Support commands up to 512 bytes in length

This is convenient for some larger CMD_SEND_BYTES payloads, but not strictly required. Since we have space in the Arduino memory, might as well use it.

* Introduce new CMD_SEND_BYTES

Create a new CMD_SEND_BYTES and corresponding response, which can be used to send multiple bytes in a single command. This reduces command/response overhead, especially for larger numbers of bytes.

* Start using new CMD_SEND_BYTES from C#

The previous commit introduced CMD_SEND_BYTES to the firmware, now let's start using it. This is done by adding a new function to use the command, and transitioning SendData(data) to use it.

* Update copyrights

* Add WRITE_TRANSFER_SPEED command

Add support for a new command WRITE_TRANSFER_SPEED, which allows the independent setting of tx and rx transfer speeds in kilobaud.

This is useful for some manufacturer-specific features such as high-speed keyloading. Currently not used by the software, being added for future use.

* Fix one remaining version

* Add version-based support for new commands

We don't want to break backwards compatibility with the existing KFDshields using the current adapter protocol version, so add some feature flags that enable the new commands if the adapter FW has already been bumped up. This lets us gracefully fail to existing commands if the KFD hardware is still on old FW.

* Make exceptions for timeout more specific

* Add debug logging of bytes send/received