Fix multiple keyloading bug
Multiple keyloading has been broken for large numbers of keys since the introduction of `CMD_SEND_BYTES`. This is due to us overrunning a buffer on the adapter when too many bytes are sent to it when certain bytes have to be escaped in the message.

Reduce our maximum message size for `CMD_SEND_BYTES` to handle even the worst-case situation where *every* byte has to be escaped.

Tested on a KFDnano with 20 AES-256 keys with value `6161616161616161616161616161616161616161616161616161616161616161`. Observed that they loaded correctly, and the maximum message size logged was ~450 bytes.

Thanks to @duggerd for pointers on where to look for this.