Add support for key sharing to KVL 4000 (#19)
This commit introduces support for key sharing from KFDtool software to the Motorola KVL 4000.

Sharing is implemented as an automatic feature when using Keyload or Multiple Keyload functionality; the software detects whether the connected device is a KVL, and adjusts its behaviour appropriately.

The main differences are a new opcode (`D1` indicating a KVL vs `D0` indicating an MR), and limiting the KMM size for transfers to a KVL.

The latter is required because the KFD* hardware has a limited onboard buffer, and the KVL does not like when there is too much delay between bytes on the TWI data line. A limited KMM size avoids calls being split over multiple CMD_SEND_BYTES commands.