Add native-USB device type to unbreak DTR setting
The previous commit (6a5035ebf5477099179c587eeee0b1258b495dcc) unfortunately turns out to have broken support for legacy KFD* devices that rely on USB-to-serial adapters - i.e. mostly AVR-architecture devices such as the KFDshield, KFDnano, KFDmicro, and older-firmware versions of the KFDpico.

Make this new setting configurable to support new native-USB devices such as the newer-firmware version of the KFDpico and future SAMD-architecture devices.

In the process, remove the KFDmicro device type to reduce user confusion as it is simply an alias for KFDshield.