Introduce SEND_KEY_SIGNATURE_AND_READY_REQ command
This commit introduces a new `SEND_KEY_SIGNATURE_AND_READY_REQ` command to the adapter protocol, and bumps the protocol version to 2.2.0.

This command allows a key signature to be automatically followed by a `C0` ready req message.

By doing this in one command, the variable latency of the round-trip to the software is removed, which resolves some latency-sensitive bugs in the Kenwood NX and VP series radios.

At this time, the adapter firmware does not include this command; however it was tested on the Rust reference implementation that will be published prior to this PR being merged.